requirements.txt
# 核心插件系统 (Core Plugin System)
pluggy>=1.5.0

# 数据验证与建模 (Data Validation & Modeling)
pydantic>=2.0.0

# 异步I/O支持 (Asynchronous I/O Support)
asyncio-subprocess>=0.1.0

# 报告生成 (Report Generation)
jinja2>=3.1.0

# YAML配置解析 (YAML Configuration Parsing)
pyyaml>=6.0.0

# 彩色日志输出 (Colored Logging Output)
colorlog>=6.8.0

# 网络工具 (Network Utilities)
validators>=0.22.0

# 测试框架 (Testing Framework)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
